refactor(runtime): centralize AI SDK tool repair
Generated-by: Codex
